Added functionality to add audio to the database

+++ RadioPlayout/Controllers/AudioController.cs
@@ -7,12 +7,17 @@ using System.Web.Mvc;
 using PagedList;
 using System.Net;
 using System.IO;
+using System.Web.Hosting;
 
 namespace RadioPlayout.Controllers
 {
     public class AudioController : Controller
     {
 		private RadioPlayoutDb _db = new RadioPlayoutDb();
+		private readonly string uploadAudioDirectory = HostingEnvironment.MapPath("~/Content/Audio/Upload/");
+		private readonly string audioDirectory = HostingEnvironment.MapPath("~/Content/Audio/");
+		private Dictionary<string, string> errorDict = new Dictionary<string, string>();
+
 
 		// GET: Audio
 		public ViewResult Index(string sortOrder, string searchString, string currentSearchString, string audioType, string currentAudioType, string audioMinDuration, 
@@ -113,21 +118,11 @@ namespace RadioPlayout.Controllers
             return View(audio.ToPagedList(pageNumber, pageSize));
         }
 
-		[HttpPost]
-		public ActionResult AddNewAudioItem(string artistName, string audioTitle, string audioLocation, string audioDuration,
-			string audioIn, string audioOut, string audioReleaseYear, string audioType)
+		public void ValidateAudioItems(string artistName, string audioTitle, string audioLocation, string uploadPath, string audioPath, 
+			string audioDuration, string audioIn, string audioOut, string audioReleaseYear, string audioType)
 		{
-			// Placeholders when converting the timings strings to integers
-			int audioDurationInt = 0;
-			int audioInInt = 0;
-			int audioOutInt = 0;
-			int audioReleaseYearInt = 0;
-			int audioTypeInt = 0;
-			string uploadPath = Server.MapPath("~/Content/Audio/Upload/" + audioLocation);
-			string audioPath = Server.MapPath("~/Content/Audio/" + audioLocation);
-
-			// Dictionary to hold key value error pairs
-			Dictionary<string, string> errorDict = new Dictionary<string, string>();
+			// Reference to an invlid integer value
+			int integerReference;
 
 			// Validate the inputted values
 			// Validate the artist name
@@ -150,19 +145,22 @@ namespace RadioPlayout.Controllers
 				// Audio location doesn't have a value
 				errorDict.Add("AudioLocation", "The audio location cannot be empty.");
 			}
-			else if(!System.IO.File.Exists(uploadPath))
+			else if (!System.IO.File.Exists(uploadPath))
 			{
-				// Audio is not in the default upload directory
-				errorDict.Add("AudioLocation", "The audio file has not been uploaded, please try again");
+				if (!System.IO.File.Exists(audioPath))
+				{
+					// Audio is not in the /Content/Audio/Upload
+					errorDict.Add("AudioLocation", "The audio file has not been uploaded, please try again");
+				}
 			}
 
 			// Validate the audio duration
-			if (String.IsNullOrWhiteSpace(audioDuration))
+				if (String.IsNullOrWhiteSpace(audioDuration))
 			{
 				// Audio duration doesn't have a value
 				errorDict.Add("AudioDuration", "The audio duration cannot be empty.");
 			}
-			else if(!Int32.TryParse(audioDuration, out audioDurationInt))
+			else if (!Int32.TryParse(audioDuration, out integerReference))
 			{
 				// Audio Duration couldn't be converted to an integer
 				errorDict.Add("AudioDuration", "The audio duration couldn't be converted to an integer");
@@ -174,7 +172,7 @@ namespace RadioPlayout.Controllers
 				// Audio in doesn't have a value
 				errorDict.Add("AudioIn", "The audio in cannot be empty");
 			}
-			else if(!Int32.TryParse(audioIn, out audioInInt))
+			else if (!Int32.TryParse(audioIn, out integerReference))
 			{
 				// Audio in couldn't be converted to an integer
 				errorDict.Add("AudioIn", "The audio in couldn't be converted to an integer");
@@ -186,7 +184,7 @@ namespace RadioPlayout.Controllers
 				// Audio out doesn't have a value
 				errorDict.Add("AudioOut", "The audio out cannot be empty.");
 			}
-			else if(!Int32.TryParse(audioOut, out audioOutInt))
+			else if (!Int32.TryParse(audioOut, out integerReference))
 			{
 				// Audio out couldn't be converted to an integer
 				errorDict.Add("AudioOut", "The audio out couldn't be converted to an integer.");
@@ -198,7 +196,7 @@ namespace RadioPlayout.Controllers
 				// Audio release year doesn't have a value
 				errorDict.Add("AudioReleaseYear", "The audio release year cannot be empty");
 			}
-			else if(!Int32.TryParse(audioReleaseYear, out audioReleaseYearInt))
+			else if (!Int32.TryParse(audioReleaseYear, out integerReference))
 			{
 				// Audio release year couldn't be converted to an integer
 				errorDict.Add("AudioReleaseYear", "The audio release year couldn't be converted to an integer.");
@@ -210,12 +208,25 @@ namespace RadioPlayout.Controllers
 				// Audio type doesn't have a value
 				errorDict.Add("AudioType", "THe audio type cannot be empty.");
 			}
-			else if(!Int32.TryParse(audioType, out audioTypeInt))
+			else if (!Int32.TryParse(audioType, out integerReference))
 			{
 				// Audio type couldn't be converted to an integer
 				errorDict.Add("AudioType", "The audio type couldn't be converted to an integer");
 			}
+		}
 
+		[HttpPost]
+		public ActionResult AddNewAudioItem(string artistName, string audioTitle, string audioLocation, string audioDuration,
+			string audioIn, string audioOut, string audioReleaseYear, string audioType)
+		{
+			errorDict.Clear();
+
+			// Placeholders for the audio upload paths
+			string uploadPath = uploadAudioDirectory + audioLocation;
+			string audioPath = audioDirectory + audioLocation;
+
+			// Dictionary to hold key value error pairs
+			ValidateAudioItems(artistName, audioTitle, audioLocation, uploadPath, audioPath, audioDuration, audioIn, audioOut, audioReleaseYear, audioType);
 
 			if (errorDict.Count > 0)
 			{
@@ -225,12 +236,27 @@ namespace RadioPlayout.Controllers
 			}
 			else
 			{
+				// No errors were produced. The audio can be added
 				// Move the file from Content/Audio/Upload to Content/Audio
-				if (System.IO.File.Exists(audioPath))
+				if (System.IO.File.Exists(audioPath) && System.IO.File.Exists(uploadPath))
 				{
+					// Delete the file from the audio directory
 					System.IO.File.Delete(audioPath);
+					// Move the new file into the audio directory
+					System.IO.File.Move(uploadPath, audioPath);
 				}
-				System.IO.File.Move(uploadPath, audioPath);
+				else if(!System.IO.File.Exists(audioPath) && System.IO.File.Exists(uploadPath))
+				{
+					// Move the file from the upload directory to the audio directory
+					System.IO.File.Move(uploadPath, audioPath);
+				}
+				else
+				{
+					// When the user is adding a new audio file this part of the code should next run.
+					// This would only run if the audio is present in /Content/Audio but not in the /Content/Audio/Upload directory
+					Response.StatusCode = (int)HttpStatusCode.BadRequest;
+					errorDict.Add("Audio", "There wasn an error with the file upload. Please try again");
+				}
 
 				// Remove all the files in the Upload directory
 				System.IO.DirectoryInfo di = new DirectoryInfo(Server.MapPath("~/Content/Audio/Upload/"));
@@ -240,18 +266,18 @@ namespace RadioPlayout.Controllers
 				}
 
 				// Get the audio type id
+				int audioTypeInt = Int32.Parse(audioType);
 				AudioType audioTypeRef = _db.AudioType.Where(r => r.AudioTypeId.Equals(audioTypeInt)).First();
 
-				// No errors were produced. The audio can be added
 				Audio newAudio = new Audio
 				{
 					ArtistName = artistName,
 					AudioTitle = audioTitle,
 					AudioLocation = "Content/Audio/" + audioLocation,
-					AudioDuration = audioDurationInt,
-					AudioIn = audioInInt,
-					AudioOut = audioOutInt,
-					AudioReleaseYear = audioReleaseYearInt,
+					AudioDuration = Int32.Parse(audioDuration),
+					AudioIn = Int32.Parse(audioIn),
+					AudioOut = Int32.Parse(audioOut),
+					AudioReleaseYear = Int32.Parse(audioReleaseYear),
 					AudioType = audioTypeRef
 				};
 
@@ -264,24 +290,112 @@ namespace RadioPlayout.Controllers
 		}
 
 		[HttpPost]
-		public ActionResult UploadAudioFile()
+		public ActionResult UpdateAudioItem(string audioId, string artistName, string audioTitle, string audioLocation, string audioDuration,
+			string audioIn, string audioOut, string audioReleaseYear, string audioType)
 		{
-			HttpPostedFileBase file = Request.Files[0];
-			int fileSize = file.ContentLength;
-			string fileName = file.FileName;
-			string mimeType = file.ContentType;
-			string filePath = "/Content/Audio/Upload/" + fileName;
+			errorDict.Clear();
 
-			file.SaveAs(Server.MapPath(filePath));
+			// Placeholders for the audio upload paths
+			string uploadPath = uploadAudioDirectory + audioLocation;
+			string audioPath = audioDirectory + audioLocation;
 
-			return Json(filePath);
+			// Dictionary to hold key value error pairs
+			ValidateAudioItems(artistName, audioTitle, audioLocation, uploadPath, audioPath, audioDuration, audioIn, audioOut, audioReleaseYear, audioType);
+
+			// Check whether the audioID exists
+			// Placeholder for the audioIDInt
+			int audioIdInt = 0;
+			if (String.IsNullOrWhiteSpace(audioId))
+			{
+				errorDict.Add("AudioItem", "There was an error updating this audio item. Please try again.");
+			}
+			else if(Int32.TryParse(audioId, out audioIdInt))
+			{
+				errorDict.Add("AudioItem", "There wasn an error updating this audio item. Please try again");
+			}
+
+			if (errorDict.Count > 0)
+			{
+				// The validation brought up errors. Do NOT Update the audio item
+				Response.StatusCode = (int)HttpStatusCode.BadRequest;
+				return Json(errorDict);
+			}
+			else
+			{
+				// Move the file from /Content/Audio/Upload to Content/Audio
+				if (System.IO.File.Exists(uploadPath))
+				{
+					// If the file is in uploadPath. Check whether it exists in the /Content/Audio directory
+					if (System.IO.File.Exists(audioPath))
+					{
+						System.IO.File.Delete(uploadPath);
+					}
+
+					System.IO.File.Move(uploadPath, audioPath);
+				}
+
+				// Remove all the files in the Upload directory
+				System.IO.DirectoryInfo di = new DirectoryInfo(Server.MapPath("~/Content/Audio/Upload/"));
+				foreach (FileInfo file in di.GetFiles())
+				{
+					file.Delete();
+				}
+
+				// Get the audio type id
+				AudioType audioTypeRef = _db.AudioType.Where(r => r.AudioTypeId.Equals(Int32.Parse(audioType))).First();
+
+				var audioItems = _db.Audio.Where(r => r.AudioId.Equals(audioIdInt));
+				foreach(Audio audioItem in audioItems)
+				{
+					audioItem.ArtistName = artistName;
+					audioItem.AudioTitle = audioTitle;
+					audioItem.AudioLocation = audioPath;
+					audioItem.AudioDuration = Int32.Parse(audioDuration);
+					audioItem.AudioIn = Int32.Parse(audioIn);
+					audioItem.AudioOut = Int32.Parse(audioOut);
+					audioItem.AudioReleaseYear = Int32.Parse(audioReleaseYear);
+					audioItem.AudioType.AudioTypeId = audioTypeRef.AudioTypeId;
+				}
+				_db.SaveChanges();
+
+				return Json("Success");
+			}
 		}
 
 		[HttpPost]
-		public ActionResult UpdateAudioItem()
+		public ActionResult UploadAudioFile()
 		{
-			return Json("", JsonRequestBehavior.AllowGet);
+			if (Request.Files.Count > 0)
+			{
+				HttpPostedFileBase file = Request.Files[0];
+				int fileSize = file.ContentLength;
+				string fileName = file.FileName;
+				string mimeType = file.ContentType;
+				string filePath = "/Content/Audio/Upload/" + fileName;
+
+				file.SaveAs(Server.MapPath(filePath));
+
+				return Json(filePath);
+			}
+			else
+			{
+				Response.StatusCode = (int)HttpStatusCode.BadRequest;
+				return Json("Invalid File Upload");
+			}
 		}
 
+		[HttpPost]
+		public ActionResult GetAudioItemInfo(string audioId)
+		{
+			int audioIdInt = 0;
+			if(!Int32.TryParse(audioId, out audioIdInt))
+			{
+				errorDict.Add("Audio", "There wasn an error with the file upload. Please try again");
+			}
+
+			var audioItem = _db.Audio.Where(r => r.AudioId.Equals(audioIdInt));
+
+			return Json(audioItem);
+		}
 	}
 }
